requirements.txt
attrs==24.3.0
bottle==0.13.2
certifi==2024.12.14
charset-normalizer==3.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
playwright==1.41.1
pluggy==1.5.0
proxy_tools==0.1.0
pyaudio==0.2.14
pyee==11.0.1
PyQt5==5.15.11
PyQtWebEngine==5.15.7
PySocks==1.7.1
pystray==0.19.5
pytest==8.3.4
python-dotenv==1.0.1
python-xlib==0.33
pywebview==5.3.2
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
vosk==0.3.45
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
rumps==0.3.0
